feat: NewTask component
